adding start product page
