--- src/components/App/App.scss
@@ -1,8 +1,0 @@
-body {
-  font-family: sans-serif;
-}
-
-.header {
-  padding: 10px;
-  margin: 10px;
-}
