+++ src/components/ProductListView/ProductListView.css
@@ -1,0 +1,19 @@
+.phone-list {
+  border: 1px solid grey;
+  border-radius: 5px;
+  display: flex;
+  margin-bottom: 10px;
+  padding: 15px 4px 4px 4px;
+}
+.picphone {
+  display: block;
+  max-width: 100px;
+  height: auto;
+}
+.phone-text {
+  margin-left: 14px;
+}
+.description {
+  font-family: 'Courier New', Courier, monospace, sans-serif;
+  font-size: 12px;
+}
